DATACMNS-935 - Removed dependency to Spring's Descriptor.
The Descriptor class is nested in the ReactiveAdapter interface which in turn has a strong dependency to Project Reactor as it exposes some if its types as method return types. That means we can't refer to Spring's Descriptor without Project Reactor on the classpath.

Introduced a custom Descriptor that exposes the setup needs we have using a fluent API.

We now also only conditionally actually the ReactiveAdapterRegistry as eagerly instantiating it in RectiveWrapperConverter causes a strong dependency on Reactor.

Related tickets: DATACMNS-836, SPR-14902.